* 'CSSyBorrado' of https://github.com/ErickRosasMendoza/SIBCSS:
  Tutorial completado: Validacion , Lista de Admins, Descargas Admin bien y Delete Gral.
  Validacion de Alumnos en lado Admin, Informacion completa en cada tabla, Nuevos CSS para <th> y Borrado codigo basura.
  Borrando comentarios en CSS de Login
  Borrando linea comentada del CSS del Login
  Se agrego info al login, se arreglaron los mensajes, Validacion de Boleta y Delete en Cascada (Alumno y Admin)
  Cambios en el login y Borrado en Cascada, Cookies y Boleta Duplicada

+++ front/src/components/AdminBajaArchivos.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Redirect,Link } from 'react-router-dom';
+import {Link} from 'react-router-dom';
 import axios from 'axios';
 import BorrarDoc from './BorrarDoc';
 import ActualizarComentario from './ActualizarComentario';
@@ -10,6 +10,7 @@ class AdminBajaArchivos extends React.Component {
 
     estadoRef = React.createRef();
     comentarioRef=React.createRef();
+    eliminarRef = React.createRef();
 
     state = {
         idAlumno: this.props.id,
@@ -28,6 +29,7 @@ class AdminBajaArchivos extends React.Component {
         cambioEstado: {},
         statusBaja: false,
         statusEstado: null,
+        statusEliminar: null
     };
 
     changeState = () =>{
@@ -102,12 +104,33 @@ class AdminBajaArchivos extends React.Component {
             });
     }//Fin de getAlumno()
 
-    deleteTipoBaja = () => {
-        axios.delete("solicitudBaja/delete/"+this.props.id)
-        .then(res => {
-            window.location.href = "./" + this.props.id
+    statusDelete = () => {
+        this.setState({
+            statusEliminar: this.eliminarRef.current.value
         })
-    }//Fin de deleteTipoBaja
+    }//Fin de Status Delete
+
+    deleteTipoBaja = () => {
+        if(this.state.statusEliminar === "true"){
+            try{
+                axios.delete("solicitudBaja/delete/"+this.props.id)
+                .then(res => {
+                window.location.reload()
+            });
+            }
+            finally{
+                this.setState({
+                    statusEliminar: "false"
+                })
+            }//Fin de finally
+        }
+    else{
+        this.setState({
+            statusEliminar: "false"
+        });
+        }//Fin de else
+     
+}//Fin de deleteTipoBaja
 
     cancelComentario = () => {
         this.setState({
@@ -134,8 +157,6 @@ class AdminBajaArchivos extends React.Component {
         }
     }//Fin de Cambiar Estado
 
-
-
     fileChange = (event) => {
        this.setState({
             file: event.target.files[0]
@@ -161,7 +182,7 @@ class AdminBajaArchivos extends React.Component {
     }
 
     upLoad = () => {
-        if(this.state.file && this.state.file != null && this.state.file != undefined){
+        if(this.state.file && this.state.file !== null && this.state.file !== undefined){
             const fd = new FormData();
             console.log(this.state);
             fd.append('file', this.state.file, this.state.file.name)
@@ -256,7 +277,7 @@ class AdminBajaArchivos extends React.Component {
                                     <strong>Estado:</strong>{this.state.Baja.estado}
                                 </div>
                              
-                                <strong>cambiar estado de la revision</strong>
+                                <strong>Cambiar estado de la revisión:</strong>
                                 <div className="center">
                                     <select name="estado" ref={this.estadoRef} onChange={this.changeState}>
                                          <option value=""></option>
@@ -269,16 +290,39 @@ class AdminBajaArchivos extends React.Component {
                                     <br />
                                 </div>
                                 <br />
-                           {/* <button id="btn_deleteRegistro" onClick={this.deleteDictamen}>Borrar Registro</button> */}
+                                <strong>¿Eliminar registro?</strong>
+                                <div className="center">
+                                    <select name="eliminar" ref={this.eliminarRef} onChange={this.statusDelete}>
+                                        <option value="false">NO</option>
+                                        <option value="true">SI</option>
+                                    </select>
+                                    {(() => {
+                                        switch(this.state.statusEliminar){   
+                                            case "false":
+                                            return (
+                                            <a className="warning_search">¡Seleccione "SI" para eliminar alumno!</a>
+                                            );
+                                            break;
+                                            default:
+                                                return(
+                                                    <a className="warning_search">¡Por seguridad antes de eliminar el registro, no debe tener documentos almacenados!</a>
+                                                )
+                                                break;
+                                        }
+                                        })()}
+                                    <button id="btn_delete" onClick={this.deleteTipoBaja}>Eliminar</button>
+                                    <br />
+                                </div>
                             </div>
                             </div>
                         </div>{/**fincontenedor */}
-                        <br />
-                        <br />  
+                        <br /> 
+                        <br /> 
+                        <table>
                                 <tbody>
                                         <tr>
-                                            <td className="table_lista"><strong>Documentos</strong></td>
-                                            <td className="table_lista"><strong>Comentario</strong></td>
+                                            <td className="table_lista, table_title"><strong>Documentos</strong></td>
+                                            <td className="table_lista, table_title"><strong>Comentario</strong></td>
                                         </tr>
                                     </tbody>
                                     {this.state.listar.map((lista1, i) =>
@@ -306,6 +350,7 @@ class AdminBajaArchivos extends React.Component {
                                             </tr>
                                     </tbody>
                                     )}
+                            </table>
                                     <br />
                                     <div  className="archivosAdminCenter" ><strong>Enviar archivo PDF</strong></div> <br />  
                                     <input type="file" name = "file" onChange={this.fileChange} />
